requirements.txt
django==1.11.3
djangorestframework>=3.6.3
mysql-connector==2.1.4
mysqlclient==1.3.10
django-cors-headers==2.1.0
